requirements.txt
cffi==1.14.5
cryptography==3.4.7
DateTime==4.3
pycparser==2.20
pyOpenSSL==20.0.1
python-nmap==0.6.4
pytz==2021.1
six==1.15.0
zope.interface==5.4.0
